+++ R/gg_bin2d.R
@@ -67,7 +67,8 @@
 #'   gg_bin2d(
 #'     x = carat,
 #'     y = price,
-#'     pal = viridis::cividis(9)
+#'     pal = viridis::cividis(9),
+#'     y_limits = c(NA, 20000)
 #'   )
 #'
 gg_bin2d <- function(
@@ -439,7 +440,7 @@ gg_bin2d <- function(
   #Get the positional scales right first
   if (stat != "sf") {
     if (rlang::is_null(x_limits)) {
-      if (stat %in% c("bin", "bin2d", "bin_2d", "binhex")) {
+      if (stat == "bin") {
         if (x_numeric) x_limits <- c(NA, NA)
         if (x_date | x_datetime | x_time) {
           x_limits <- c(lubridate::NA_Date_, lubridate::NA_Date_)
@@ -474,7 +475,7 @@ gg_bin2d <- function(
     }
 
     if (rlang::is_null(y_limits)) {
-      if (stat %in% c("bin", "bin2d", "bin_2d", "binhex")) {
+      if (stat == "bin") {
         if (y_numeric) y_limits <- c(NA, NA)
         if (y_date | y_datetime | y_time) {
           y_limits <- c(lubridate::NA_Date_, lubridate::NA_Date_)