+++ R/gg_hex.R
@@ -71,7 +71,9 @@
 #'     x = carat,
 #'     y = price,
 #'     pal = viridis::cividis(9),
-#'     coord = coord_cartesian(clip = "on"))
+#'     y_limits = c(NA, 20000),
+#'     coord = coord_cartesian(clip = "on")
+#'   )
 #'
 gg_hex <- function(
     data = NULL,
@@ -442,7 +444,7 @@ gg_hex <- function(
   #Get the positional scales right first
   if (stat != "sf") {
     if (rlang::is_null(x_limits)) {
-      if (stat %in% c("bin", "bin2d", "bin_2d", "binhex")) {
+      if (stat == "bin") {
         if (x_numeric) x_limits <- c(NA, NA)
         if (x_date | x_datetime | x_time) {
           x_limits <- c(lubridate::NA_Date_, lubridate::NA_Date_)
@@ -477,7 +479,7 @@ gg_hex <- function(
     }
 
     if (rlang::is_null(y_limits)) {
-      if (stat %in% c("bin", "bin2d", "bin_2d", "binhex")) {
+      if (stat == "bin") {
         if (y_numeric) y_limits <- c(NA, NA)
         if (y_date | y_datetime | y_time) {
           y_limits <- c(lubridate::NA_Date_, lubridate::NA_Date_)